Images can now be resized with a double value (casts back to an int) to allow for smaller/larger images based off of a base image, and fixed the weird issue with addImage by coordinating the Queues better.
Problem was in the time it took for the program to retrieve the image to make a DrawnImage object, it having added the name to the Queue polled the name before the DrawnImage was added to the ElementQueue. Lots of concurrency issues I need to figure out.

+++ visual/panel/element/DrawnImage.java
@@ -9,7 +9,7 @@ public class DrawnImage extends Element{
 //---  Instance Variables   -------------------------------------------------------------------
 	
 	/** */
-	private int scaleFactor;
+	private double scaleFactor;
 	/** */
 	private Image image;
 	
@@ -25,7 +25,7 @@ public class DrawnImage extends Element{
 	 * @param scale
 	 */
 	
-	public DrawnImage(int x, int y, int prior, boolean inCenter, Image img, int scale) {
+	public DrawnImage(int x, int y, int prior, boolean inCenter, Image img, double scale) {
 		setX(x);
 		setY(y);
 		image = img;
@@ -54,7 +54,7 @@ public class DrawnImage extends Element{
 	
 	public void drawToScreen(Graphics g) {
 		Toolkit tk = Toolkit.getDefaultToolkit();
-		Image drawImage = image.getScaledInstance(image.getWidth(null) * scaleFactor, image.getHeight(null) * scaleFactor, Image.SCALE_DEFAULT);
+		Image drawImage = image.getScaledInstance((int)(image.getWidth(null) * scaleFactor), (int)(image.getHeight(null) * scaleFactor), Image.SCALE_DEFAULT);
 		while(!tk.prepareImage(drawImage, -1, -1, null)){	}
 		g.drawImage(drawImage, getX() - (center ? drawImage.getWidth(null)/2 : 0), getY() - (center ? drawImage.getHeight(null)/2 : 0), null);
 	}
